Register sonos entity services in async_setup
